requirements.txt
pillow==10.4.0
pycryptodome==3.20.0
pypng==0.20220715.0
qrcode==7.4.2
typing_extensions==4.12.2
